Rename ext.py to limiter.py, and also change some formatting
